separate logging from executor.go into log.go
